Safari-609.1.11.1/LayoutTests/fast/dom/attributes-configurable.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
description("Checks that attributes are on the prototype are configurable.");

var event = new Event("test");
shouldBe("event.eventPhase", "0");
shouldBeTrue("event.__proto__.hasOwnProperty('eventPhase')");
var properties = Object.getOwnPropertyDescriptor(event.__proto__, "eventPhase");
shouldBeTrue("properties.configurable");
shouldBeTrue("properties.enumerable");
shouldBeTrue("delete event.__proto__.eventPhase");
shouldBe("event.eventPhase", "undefined");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
